requirements.txt
openai>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
chromadb>=0.4.0
requests>=2.31.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pypdf>=4.0.0
pdfplumber>=0.11.0
icalendar>=5.0.0
flask>=3.0.0
flask-cors>=4.0.0
